requirements.txt
absl-py==2.3.1
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
apache-tvm-ffi==0.1.9
astor==0.8.1
asttokens @ file:///croot/asttokens_1743630435401/work
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.44.1
blake3==1.0.8
blis==1.3.3
cachetools==7.0.2
catalogue==2.0.10
cbor2==5.8.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.9.0
comm @ file:///croot/comm_1709322850197/work
compressed-tensors==0.13.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==13.1.1
cuda-pathfinder==1.4.0
cuda-python==13.1.1
cuda-toolkit==13.0.2
cupy-cuda12x==14.0.1
cycler==0.12.1
cymem==2.0.13
DataProperty==1.1.0
datasets==3.0.1
debugpy @ file:///home/task_175706710642905/conda-bld/debugpy_1757067121571/work
decorator @ file:///home/task_175734035162575/conda-bld/decorator_1757341232127/work
deepspeed==0.15.4
depyf==0.20.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
evaluate==0.4.6
executing @ file:///home/task_175706124557094/conda-bld/executing_1757061259817/work
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
filelock==3.20.0
flash_attn==2.8.3
flashinfer-python==0.6.3
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
gguf==0.18.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-genai==1.57.0
grpcio==1.78.0
grpcio-reflection==1.78.0
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hydra-colorlog==1.2.0
hydra-core==1.3.0
idna==3.10
ijson==3.5.0
interegular==0.3.3
ipykernel @ file:///home/task_175758336750008/conda-bld/ipykernel_1757583382803/work
ipython @ file:///home/task_175932809916977/conda-bld/ipython_1759328496611/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1751991368470/work
kiwisolver==1.4.9
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lm_eval==0.4.11
loguru==0.7.3
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline @ file:///home/task_175851944441289/conda-bld/matplotlib-inline_1758520101789/work
mbstrdecoder==1.1.4
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.9.1
model-hosting-container-standards==0.1.13
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.15
mwparserfromhell==0.7.2
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numba==0.61.2
numexpr==2.13.1
numpy==2.2.3
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cudnn-frontend==1.18.0
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-cutlass-dsl==4.4.1
nvidia-cutlass-dsl-libs-base==4.4.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.20.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvshmem-cu12==3.3.20
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==2.24.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.92
outlines_core==0.2.11
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==2.3.3
parso @ file:///croot/parso_1733963305961/work
partial-json-parser==0.2.1.1.post7
pathvalidate==3.3.1
peft==0.17.1
pexpect @ file:///croot/pexpect_1750072593878/work
pillow==12.0.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
portalocker==3.2.0
preshed==3.0.12
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
prompt_toolkit @ file:///home/task_176174483353592/conda-bld/prompt-toolkit_1761744845973/work
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/task_175706703828594/conda-bld/pure_eval_1757067053474/work
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pybind11==3.0.1
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.0
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.1
Pygments @ file:///croot/pygments_1744664109463/work
PyJWT==2.11.0
pyparsing==3.2.5
pytablewriter==1.2.1
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///home/task_175810349686595/conda-bld/pyzmq_1758103934047/work
ray==2.54.0
referencing==0.37.0
regex==2026.3.32
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rouge_score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
sacrebleu==1.5.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==5.2.0
sentencepiece==0.2.1
sentry-sdk==2.54.0
setproctitle==1.3.7
shellingham==1.5.4
six @ file:///croot/six_1744271502820/work
smart_open==7.5.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.5.2
sse-starlette==3.3.2
stack_data @ file:///home/task_175706683579810/conda-bld/stack_data_1757067036897/work
starlette==0.52.1
supervisor==4.3.0
svgling==0.5.0
svgwrite==1.4.3
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
torch==2.4.1
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.51.3
triton==3.0.0
typepy==1.3.4
typer==0.24.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.41.0
uvloop==0.22.1
wandb==0.21.4
wasabi==1.1.3
watchfiles==1.1.1
wcwidth @ file:///croot/wcwidth_1750352883074/work
weasel==0.4.3
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
Wikipedia-API==0.8.1
word2number==1.1
wrapt==2.0.1
xgrammar==0.1.29
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
